Rename gin package
